Feat: 네트워크 선택창 구현

+++ app/src/main/java/com/example/lumentree/core/model/device_connect/WiFiAPInfo.kt
@@ -2,6 +2,6 @@ package com.example.lumentree.core.model.device_connect
 
 data class WiFiAPInfo(
     val ssid: String,
-    val strength: Double,
+    val strength: Int,
     val protected: Boolean
 )
